Add contributor role

+++ app/Http/Controllers/IssuesController.php
@@ -2,19 +2,25 @@
 
 namespace App\Http\Controllers;
 
+use App\Services\CommonsService;
+use App\Services\FilesService;
 use App\Services\IssuesService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Inertia\Inertia;
 use Inertia\Response;
 
-class IssueController extends Controller
+class IssuesController extends Controller
 {
     protected $issuesService;
+    protected $filesService;
+    protected $commonsService;
 
-    public function __construct(IssuesService $issuesService)
+    public function __construct(IssuesService $issuesService, FilesService $filesService,CommonsService $commonsService)
     {
         $this->issuesService = $issuesService;
+        $this->filesService = $filesService;
+        $this->commonsService = $commonsService;
     }
 
     /**
@@ -25,8 +31,9 @@ class IssueController extends Controller
     {
         $issues_by_filter = $this->issuesService->getIssuesByCategoriesAndStatuses();
         $issues = $this->issuesService->getAllIssuesWithoutValidateStatus();
-        $categories = $this->issuesService->getCategories();
-        $statuses = $this->issuesService->getStatuses();
+
+        $categories = $this->commonsService->getCategories();
+        $statuses = $this->commonsService->getStatuses();
 
         return Inertia::render('Dashboard/index', [
             'issues' => $issues,
@@ -56,7 +63,7 @@ class IssueController extends Controller
     public function report_a_bug(): Response
     {
         return Inertia::render('ReportABug/index', [
-            'categories' => $this->issuesService->getCategories()
+            'categories' => $this->commonsService->getCategories()
         ]);
     }
 
@@ -67,7 +74,7 @@ class IssueController extends Controller
     public function feature_request(): Response
     {
         return Inertia::render('FeatureRequest/index', [
-            'categories' => $this->issuesService->getCategories()
+            'categories' => $this->commonsService->getCategories()
         ]);
     }
 
@@ -76,7 +83,7 @@ class IssueController extends Controller
     {
         $issue = $this->issuesService->getById($id);
 
-        if($issue->type->name ==Config::get('constants.types.feature')){
+        if($issue->type->name == Config::get('constants.types.feature')){
             return Inertia::render('404/index');
         }
         return Inertia::render('Issue/display', ['issue' => $issue]);
@@ -96,15 +103,19 @@ class IssueController extends Controller
 
     public function display_edit_issue($id): Response
     {
-        $issue = $this->issuesService->getById($id);
-
-        if($issue->type->name ==Config::get('constants.types.feature')){
+        if(!$this->commonsService->isManager()) {
             return Inertia::render('404/index');
         }
 
-        $categories = $this->issuesService->getCategories();
-        $statuses = $this->issuesService->getStatuses();
-        $types = $this->issuesService->getTypes();
+        $issue = $this->issuesService->getById($id);
+
+        if($issue->type->name == Config::get('constants.types.feature')){
+            return Inertia::render('404/index');
+        }
+
+        $categories = $this->commonsService->getCategories();
+        $statuses = $this->commonsService->getStatuses();
+        $types = $this->commonsService->getTypes();
 
         return Inertia::render('Issue/edit', [
             'issue' => $issue,
@@ -116,6 +127,10 @@ class IssueController extends Controller
 
     public function display_edit_feature($id): Response
     {
+        if(!$this->commonsService->isManager()) {
+            return Inertia::render('404/index');
+        }
+
         $issue = $this->issuesService->getById($id);
 
 
@@ -123,9 +138,9 @@ class IssueController extends Controller
             return Inertia::render('404/index');
         }
 
-        $categories = $this->issuesService->getCategories();
-        $statuses = $this->issuesService->getStatuses();
-        $types = $this->issuesService->getTypes();
+        $categories = $this->commonsService->getCategories();
+        $statuses = $this->commonsService->getStatuses();
+        $types = $this->commonsService->getTypes();
 
         return Inertia::render('Feature/edit', [
             'feature' => $issue,
@@ -134,9 +149,6 @@ class IssueController extends Controller
             'types' => $types,
         ]);
     }
-
-
-
 
     /**
     * Create an issue
@@ -148,9 +160,6 @@ class IssueController extends Controller
     {
         $messages = [
             "files.max" => "Maximum amount of files authorized is: 4",
-            "files.*.mimes" => "File type unauthorized Only jpg,jpeg,png,log,txt and gifs",
-            "files.*.max" => "File too big, maximum allowed is 2MB/file",
-            "files.*.size" => "File too big, maximum allowed is 2MB/file",
         ];
         request()->validate([
             'description' => ['required'],
@@ -159,15 +168,18 @@ class IssueController extends Controller
             'steps_to_reproduce' => ['required'],
             'user_discord_id' => ['required'],
             'category_id' => ['required'],
-            'files.*' => ['mimes:jpg,jpeg,png,log,txt,gif|max:2000'],
             'files' => ['max:4'],
         ], $messages);
 
         $data = $request->post();
+
         if($request->file('files')){
+            $this->filesService->validate_many($request->file('files'));
+
             $data['files'] = $request->file('files');
         }
-        $bug_type = $this->issuesService->getTypeByName(Config::get('constants.types.bug'));
+
+        $bug_type = $this->commonsService->getTypeByName(Config::get('constants.types.bug'));
         $data['type_id'] = $bug_type->id;
 
         $issue = $this->issuesService->create($data);
@@ -182,23 +194,23 @@ class IssueController extends Controller
     {
         $messages = [
             "files.max" => "Maximum amount of files authorized is: 4",
-            "files.*.mimes" => "File type unauthorized Only jpg,jpeg,png,log,txt and gifs",
-            "files.*.max" => "File too big, maximum allowed is 2MB/file",
-            "files.*.size" => "File too big, maximum allowed is 2MB/file",
         ];
+
         request()->validate([
             'description' => ['required'],
             'category_id' => ['required'],
-            'files.*' => ['mimes:jpg,jpeg,png,log,txt,gif|max:2000'],
             'files' => ['max:4'],
         ], $messages);
 
         $data = $request->post();
         if($request->file('files')){
+            $this->filesService->validate_many($request->file('files'));
+
             $data['files'] = $request->file('files');
         }
 
-        $feature_type = $this->issuesService->getTypeByName(Config::get('constants.types.feature'));
+
+        $feature_type = $this->commonsService->getTypeByName(Config::get('constants.types.feature'));
         $data['type_id'] = $feature_type->id;
 
         $issue = $this->issuesService->create($data);
@@ -216,6 +228,10 @@ class IssueController extends Controller
     */
     public function edit(Request $request)
     {
+        if(!$this->commonsService->isManager()) {
+            return response('Unauthorized',401);
+        }
+
         request()->validate([
             'description' => ['required'],
             'category_id' => ['required'],
@@ -234,38 +250,10 @@ class IssueController extends Controller
 
     public function delete($id)
     {
-        $this->issuesService->delete($id);
-    }
-
-    public function download_file($id)
-    {
-       return $this->issuesService->downloadFile($id);
-    }
-
-    public function delete_file($id)
-    {
-        return $this->issuesService->deleteFile($id);
-    }
-
-    public function add_single_file(Request $request)
-    {
-        $messages = [
-            "files.max" => "Maximum amount of files authorized is: 4",
-            "files.*.mimes" => "File type unauthorized Only jpg,jpeg,png,log,txt and gifs",
-            "files.*.max" => "File too big, maximum allowed is 2MB/file",
-            "files.*.size" => "File too big, maximum allowed is 2MB/file",
-        ];
-
-        request()->validate([
-            'file.*' => ['mimes:jpg,jpeg,png,log,txt,gif|max:2000'],
-            'files' => ['max:1'],
-        ], $messages);
-
-        if(!$request->file('file')){
-            return response('No files submitted', 400);
+        if(!$this->commonsService->isManager()) {
+            return response('Unauthorized',401);
         }
-        $fileArray = $this->issuesService->storeFiles([$request->file('file')], $request->issue_id);
 
-        return response($fileArray[0]);
+        return $this->issuesService->delete($id);
     }
 }
